+++ index.php
@@ -59,7 +59,7 @@
                 <div class="col d-flex justify-content-end">
                     <div class="row">
                         <div class="col">
-                            <p class="text-center" id="version_text" style="font-weight: normal;padding-top: 12px;color: rgb(0,0,0);">CP 4.3.2a<br><a href="https://github.com/magdesign/PocketVJ-CP-exh/issues" style="text-align: right;" target="_blank"><img class="help_logo" src="assets/img/help_issue.svg" style="text-align: right;padding: 8px;max-width: 60%;width: 40px;" data-toggle="tooltip" title="issues"></a><a href="/docs/html/index.html" style="text-align: right;" target="_blank"><img class="help_logo" src="assets/img/help.svg" style="text-align: right;padding: 8px;max-width: 60%;width: 40px;" data-toggle="tooltip" title="read the f**king docs!"></a><br></p>
+                            <p class="text-center" id="version_text" style="font-weight: normal;padding-top: 12px;color: rgb(0,0,0);">CP 4.3.3<br><a href="https://github.com/magdesign/PocketVJ-CP-exh/issues" style="text-align: right;" target="_blank"><img class="help_logo" src="assets/img/help_issue.svg" style="text-align: right;padding: 8px;max-width: 60%;width: 40px;" data-toggle="tooltip" title="issues"></a><a href="/docs/html/index.html" style="text-align: right;" target="_blank"><img class="help_logo" src="assets/img/help.svg" style="text-align: right;padding: 8px;max-width: 60%;width: 40px;" data-toggle="tooltip" title="read the f**king docs!"></a><br></p>
                         </div>
                     </div>
                 </div>
@@ -357,7 +357,7 @@
                                     <div class="col colforbutton"><a class="btn btn-dark action" role="button" href="?action=namefixer" data-toogle="tooltip" title="removes umlauts &amp; empty spaces">Filename Fixer</a></div>
                                     <div class="col colforbutton"><a class="btn btn-dark action" role="button" href="?action=clean" data-toggle="tooltip" title="removes . __.Trash files">Clean Hidden</a></div>
                                     <div class="col colforbutton"><a class="btn btn-dark action" role="button" href="?action=imageconform" data-toggle="tooltip" title="resize images to max. 1920x1080">Conform Images</a></div>
-                                    <div class="col colforbutton"><a class="btn btn-dark action" role="button" href="?action=mappingconverter" data-toggle="tooltip" title="import mapping from other software">Mapping Converter</a></div>
+                                    <div class="col colforbutton"></div>
                                     <div class="col colforbutton" style="height: 68px;"></div>
                                 </div>
                             </div>
@@ -557,6 +557,7 @@
                                 <div>
                                     <h3>MASK</h3>
                                     <div class="col colforbutton"><a class="btn btn-success action" role="button" href="?action=savemask">Save as mask.png</a></div>
+                                    <div class="col colforbutton"><a class="btn btn-success action" role="button" href="?action=saveoverlay">Save as overlay.png</a></div>
                                     <div class="col colforbutton">
                                         <h3>&nbsp;</h3>
                                     </div>
@@ -2679,6 +2680,7 @@ elem2.addEventListener("input", range2Value);
     <script src="assets/bootstrap/js/bootstrap.min.js"></script>
 
 
+
 <!-- Sets the "Set Time" button function--> 
  
 <script type="text/javascript">
@@ -2741,7 +2743,7 @@ setInterval(run, 3000);
         });
     });
     </script>
-    
+
 </body>
 
 </html>